refactor drivers and codegen
