+++ backend/app/api/frames.py
@@ -6,7 +6,8 @@ from flask import jsonify, request, Response
 from flask_login import login_required
 from . import api
 from app import redis
-from app.models.frame import Frame, new_frame, delete_frame, update_frame, generate_scene_nim_source
+from app.models.frame import Frame, new_frame, delete_frame, update_frame
+from app.codegen.scene_nim import write_scene_nim
 
 
 @api.route("/frames", methods=["GET"])
@@ -86,7 +87,7 @@ def api_frame_scene_source(id: int, scene: str):
     scene = [scene for scene in frame.scenes if scene.get('id') == 'default'][0]
     if not scene:
         return jsonify({'error': f'Scene {scene} not found'}), HTTPStatus.NOT_FOUND
-    return jsonify({'source': generate_scene_nim_source(frame, scene)})
+    return jsonify({'source': write_scene_nim(frame, scene)})
 
 @api.route('/frames/<int:id>/reset', methods=['POST'])
 @login_required
