+++ backend/app/codegen/utils.py
@@ -1,0 +1,3 @@
+
+def sanitize_nim_string(string: str) -> str:
+    return string.replace('\\', '\\\\').replace('"', '\\"').replace('\n', '\\n')
